Compile HelloWorld sample as a PCL
This checkin is a demonstration of how to take a concord component and compile it as a Portable Class Library (PCL) project. A PCL enables you to write and build managed assemblies that can be compiled once, but run against multiple versions of the .NET Framework.

This PCL project is configured to use the target framework for .NET Framework 4.5 and Windows 8 store applications, which should enable the component to load in Visual Studio, Windows IOT core, and cross-platform .NET Core scenarios. Note that this profile is missing some framework APIs which may be useful to concord components. It may be possible to use a more full featured framework, but this is the one has been shown to work with the CoreCLR which VS 2015 will use on Windows IOT.

PCLs work because the .NET Framework team took their API which was historically defined in a few different assemblies (ex: mscorlib.dll, System.dll, etc) and refactored the API definition (the assemblies that are compiled against) into a series of 'contract' assemblies (ex: System.Collections.dll) these assemblies define the shape of the API, but aren't used at runtime. At runtime, the implementation can either be in the full version of this dll, or, and this is always the case when running against the desktop CLR, the API can still be implemented in the traditional assembly (ex:mscorlib). In this later case there is a 'façade' version of the assembly that just forwards all the types back to the traditional assembly. There is a longer description of PCLs in a blog post from the .NET team: http://blogs.msdn.com/b/dotnet/archive/2014/12/04/introducing-net-core.aspx