VS 2017 support part 1 (#34)
With this commit, the "Hello World" projects should now work in VS 2017 RC3 (and probably older versions too).
This is also the first part of support in the IrisExtension. But the fix for that one is more complicated as there are breaking changes in System.Reflection.Metadata. So I don't know if it will even be possible to use the same expression compiler in both (unless of course it didn't use System.Reflection.Metadata).
This also updates the project order in the .sln files so that the
default startup project is the .vsix project.