Update samples to use the 16.0 GA Nuget packages (#44)
The finial 16.0 GA nuget packages for Concord extensibility are now on nuget.org. This updates the samples to consume this version.

Note that the GA version of the Microsoft.VSSDK.Debugger.VSDebugEng package has a slightly different directory layout to make the Nuget package manager happy. This update these projects to expect this new layout.